Update Scintilla to 2020-03-29 8098:2228fb724ce7.

+++ scintilla/lexlib/PropSetSimple.h
@@ -17,7 +17,7 @@ public:
 	void Set(const char *key, const char *val, size_t lenKey = -1, size_t lenVal = -1);
 	void SetMultiple(const char *);
 	const char *Get(const char *key) const;
-	int GetExpanded(const char *key, char *result) const;
+	size_t GetExpanded(const char *key, char *result) const;
 	int GetInt(const char *key, int defaultValue = 0) const;
 };
